Small signature fix

+++ include/rf/render.h
@@ -157,7 +157,7 @@ uint32          MakeVertexArrayObject();
 uint32          AddIBO(uint32 Usage, size_t Size, void const *Data);
 uint32          AddEmptyVBO(size_t Size, uint32 Usage);
 void            FillVBO(uint32 Attrib, uint32 AttribStride, uint32 Type, size_t ByteOffset, size_t Size, void const *Data);
-void            UpdateVBO(uint32 VBO, size_t ByteOffset, uint32 Size, void *Data);
+void            UpdateVBO(uint32 VBO, size_t ByteOffset, size_t Size, void *Data);
 void            DestroyMesh(mesh *Mesh);
 void            RenderMesh(mesh *Mesh);
